+++ flask_dance/consumer/__init__.py
@@ -1,4 +1,4 @@
 from .oauth1 import OAuth1ConsumerBlueprint
 from .oauth2 import OAuth2ConsumerBlueprint
-from .base import oauth_authorized
+from .base import oauth_authorized, oauth_error